requirements.txt
# Core dependencies
openai>=1.0.0
python-pptx>=0.6.21
python-dotenv>=0.19.0
Pillow>=8.0.0
numpy>=1.21.0
PyMuPDF>=1.20.0

# Computer Vision (for visual element detection)
opencv-python>=4.5.0

# Machine Learning (for text analysis)
scikit-learn>=1.0.0

# Performance monitoring
psutil>=5.8.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0